# Address performance issues on section splitting