Hi
The current parsing that splits the configuration from the device into sections is very slow. Running on a `DCS-7308-F` (with ~200 interfaces) this takes around **90 seconds**.

There are 2 reasons for that:
1- calling `getall()` on `api('switchports')` and `api('interfaces')` calls this twice
2- the regex approach is fairly slow, as it parses the whole config _for each interface_ (exacerbated by the first issue).

With this patch, the config is parsed once into sections. Looking at a flamegraph, it can still be improved however the impact on surrounding code would be a lot larger and the benefits are not that big (passing compiled objects or not compiling exact key matches like `interface Ethernet1`)

With this change applied, the parsing execution goes down to **2 seconds**.

The total time of execution for my device went from ~150 seconds to 27 seconds.
Further optimization by receiving an already compiled regex object when necessary (and a string for exact match when not) would bring this to 23 seconds.

This is important for me because I'm handling a fleet of 260 Arista devices and it adds up quite a bit when doing big changes.