@dlyssenko 2022-10-15
Hi @DavidVentura, nice handling of the section splitting. However, there are side effects. Say, the config section is this:
```
section foo
  bar
  baz
  qux
```
with the original code, `section( 'baz' )` is still going to find that section (even though the match is in the mid of the section itself) and the output would be fragmented (it'd be only from the point of the match), while with your handling, only matching on any of the section line (`section foo`) would find that section.

Idiomatically, I think your handling is better, but it might break the existing customer scripts who might rely on that original quirky behavior. 

Let me check with my colleagues on the possible collaterals before allowing to merge your code.