@dlyssenko 2022-10-18
After discussion with @mharista  we decided to allow this PR, the deviation in behavior from the original behavior will be release noted in the next version.