@dlyssenko 2022-10-18
hi @DavidVentura - could you please test PR #245 - I enhanced `section()` call to match on the sub-sections too, so wonder if from performance point of view if it's not too much of a slowdown?